// src/api/openapi/models/recalculation.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Order } from './order';
import { WoodList } from './woodList';
import { Expense } from './expense';
import { Paint } from './paint';
import { Lock } from './lock';
import { Workload } from './workload';


export interface Recalculation { 
    cost: number;
    expenses: Array<Expense>;
    id: number;
    km: number;
    lock: Lock;
    material_charge_percent: number;
    orders: Array<Order>;
    paints: Array<Paint>;
    pdf: string | null;
    total_sum: number;
    wood_lists: Array<WoodList>;
    workloads: Array<Workload>;
}