// src/api/openapi/model/recalculationUpdate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaintCreate } from './paintCreate';
import { ExpenseCreate } from './expenseCreate';
import { WoodListCreate } from './woodListCreate';


export interface RecalculationUpdate { 
    cost: number;
    expenses: Array<ExpenseCreate>;
    km: number;
    material_charge_percent: number;
    paints: Array<PaintCreate>;
    pdf?: string | null;
    wood_lists: Array<WoodListCreate>;
}