// src/api/openapi/model/recalculationCreate.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecalculationCreate { 
    pdf?: any | null;
    material_charge_percent: any | null;
    km: any | null;
    cost: any | null;
    expenses: any | null;
    paints: any | null;
    wood_lists: any | null;
}